Added a context menu that allows you to change the line width without going to the top of the page. Scroll wheel changes the value. Left clicking to draw or ESC closes the context menu.
Added styles for the context menu.
Made some modifications to the updateLineWidth function. It now takes a width parameter and updates both line width range inputs and previews. The range inputs have also been modified to send their own value to the updateLineWidth function when they are changed.
The line width default, min, and max values are now set from within the client.js file. Maybe we can set a cookie and save the last used size and set that as the default the next time the user plays?
Updated some comments.